feat(sys): land nvapi-macros crate + NvValue type system (Stage 1)
Introduces the `nvapi-macros` proc-macro crate and `sys/src/value.rs`
defining the NvValue<T>/NvEnum<T>/NvBits<T> repr(transparent) newtype system
+ NvValueData/NvValueEnum/NvValueBits traits. Wires it into the workspace
(members += macros, sys dep on nvapi-macros, pub mod value). No behavior
change yet — nvenum!/nvbits! still resolve to legacy macro_rules generating
bare c_int aliases; value.rs is inert until Stage 2 flips the macros.

Ported from origin/v0.2.x-macros (tip 97b9364) with three deliberate
corrections over the donor:
- compat shims: nvenum! additionally emits raw()/from_raw()/values() on the
  typed enum (signatures follow the newtype alias atomically) so hi-layer
  call sites passing alias-typed values keep compiling.
- soundness: nvbits! now emits #[repr(transparent)] on the bitflags struct
  (bitflags 1.3.2 does not add it automatically; donor's transmute was unsound).
- clippy: nvvalue_reprs! uses From<NvValue<T>> for Repr (from_over_into),
  util needless_return fixed; src/gpu.rs baseline unnecessary_cast fixed.

NvAPI_Status excluded from newtyping (stays c_int). nvversion!/NvStruct/
nv_inherit/api.rs proc-macros NOT ported (PR2 scope).

+++ src/gpu.rs
@@ -2031,7 +2031,7 @@ impl PhysicalGpu {
         let mut ctrl = unsafe {
             let b = Box::<clock::private::NV_GPU_CLOCK_CLIENT_CLK_VF_POINTS_CONTROL_PRIVATE>::new_zeroed();
             let mut b = b.assume_init();
-            b.version = NvVersion::with_version(clock::private::clk_vfp_control::MAGIC as u32);
+            b.version = NvVersion::with_version(clock::private::clk_vfp_control::MAGIC);
             b
         };
         ctrl.seed_masks_from_info(&info);
